Clear up the run_iterator, maxiter inheritance incompatibility
Remove run_iterator from AbstractSolver as to not overwrite the mixin's
implementation if AbstractSolver is inherited first.
NotImplementedError signaled that it should be implemented for all
solvers at some point anyway, which is not true.
Require maxiter to be implemented for anything that uses the mixin.
Explicitly do so in JAXopt-based stochastic solvers.
Adjust the tests.